Refact: 맵 초기화되어도 배열 유지되도록 수정
-기존의 이상현상 추출 배열을 Subsystem으로 이동
-기존의 ID들을 int32에서 uint8 형식으로 변경
-ActivateAnomaly에서 이상현상 인자를 전달하도록 수정

+++ Source/Endless_Hotel/Data/Anomaly/AnomalyData.h
@@ -11,10 +11,10 @@ struct FAnomalyData : public FTableRowBase
 
 public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 AnomalyID;
+	uint8 AnomalyID;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int32 Object_ID;
+	uint8 Object_ID;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
 	FString Anomaly_En;
